+++ plugins/gs-common/src/model/awsv1beta1/types.ts
@@ -1,0 +1,7 @@
+import { IObjectMeta } from '../metav1/types';
+
+export interface IProviderConfig {
+  apiVersion: string;
+  kind: string;
+  metadata: IObjectMeta;
+}