+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- GS plugin: Add scaffolder custom field extension for picking deployment details.
+
 ## [0.31.0] - 2024-07-30
 
 ### Changed